Refactor: code
