+++ app/src/main/java/com/example/agenda/activities/MainActivity.kt
@@ -1,14 +1,14 @@
-package com.example.agenda
+package com.example.agenda.activities
 
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
+import com.example.agenda.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+
     }
 }
