+++ app/src/main/java/com/example/agenda/activities/SplashActivity.kt
@@ -1,10 +1,11 @@
-package com.example.agenda
+package com.example.agenda.activities
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
+import com.example.agenda.R
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
