+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     android:scrollIndicators="right"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".activities.MainActivity">
