+++ app/src/main/res/layout/activity_splash.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     android:background="#F5BB00"
     android:layout_height="match_parent"
-    tools:context=".SplashActivity">
+    tools:context=".activities.SplashActivity">
 
     <TextView
         android:layout_width="match_parent"
